调整了ui，改了几个小bug

+++ WePeiYang/ecard/src/main/java/com/twt/service/ecard/model/EcardService.kt
@@ -52,7 +52,7 @@ data class EcardProfileBean(
 )
 
 data class EcardTotalConsumptionBean(
-        val total_day: Int,
+        val total_day: Double,
         val total_month: Double
 )
